preload font to avoid swap
by default, `@font-face` rules don't request the font files until the CSSOM is constructed, which is after html and css are parsed and the browser can figure out if the font is being used.

by using a link tag with preload, we can force that request to happen earlier in the process.